refactor: extract the test for GET from schema.test.ts
