+++ __test__/backend/schema/schema.test.ts
@@ -59,22 +59,9 @@ const sendRequest = (method: string, body?: Record<string, unknown>) => {
   });
 };
 
-describe("CRUD of Schema API (Using NextRequest)", () => {
+describe("CUD of Schema API (Using NextRequest)", () => {
   beforeEach(() => {
     jest.clearAllMocks();
-  });
-
-  it("should return INTERNAL_SERVER_ERROR when GET fails", async () => {
-
-    (prisma.schema.findMany as jest.Mock).mockRejectedValue(new Error("Database error"));
-
-    const request = sendRequest("GET");
-    const response = await GET(request);
-    const json = await response.json();
-
-
-    expect(response.status).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-    expect(json.error).toBe("GET Schemas: Internal Server Error");
   });
 
 
@@ -207,151 +194,5 @@ describe("CRUD of Schema API (Using NextRequest)", () => {
 
     expect(json.some((schema: Schema) => schema.name === validSchemaData.name)).toBe(false);
   });
-
-  it("should return schemas filtered by serviceIds (UUIDs)", async () => {
-    const SERVICE_ID_1 = "bd7a4c7a-1234-4c5e-b123-df12345abcd1";
-    const SERVICE_ID_2 = "bd7a4c7a-1234-4c5e-b123-df12345abcd2";
-
-    const filteredSchemas = [
-      {
-        id: 1,
-        name: "products",
-        description: "Table for product info",
-        schemaText: "CREATE TABLE products (id SERIAL PRIMARY KEY);",
-        serviceId: SERVICE_ID_1,
-      },
-      {
-        id: 2,
-        name: "users",
-        description: "Table for user info",
-        schemaText: "CREATE TABLE users (id SERIAL PRIMARY KEY);",
-        serviceId: SERVICE_ID_2,
-      },
-    ];
-
-    (prisma.schema.findMany as jest.Mock).mockImplementation(({ where }) => {
-      return Promise.resolve(
-        filteredSchemas.filter((schema) =>
-          where.AND[0].serviceId.in.includes(schema.serviceId)
-        )
-      );
-    });
-
-    const request = new NextRequest(
-      new URL(
-        `http://localhost/api/schema?serviceIds=${SERVICE_ID_1}&serviceIds=${SERVICE_ID_2}`
-      ),
-      {
-        method: "GET",
-      }
-    );
-
-    const response = await GET(request);
-    const json = await response.json();
-
-    expect(prisma.schema.findMany).toHaveBeenCalledWith({
-      where: { AND: [{ serviceId: { in: [SERVICE_ID_1, SERVICE_ID_2] } }, expect.anything()] },
-      include: {
-        service: true,
-      },
-    });
-
-    expect(response.status).toBe(StatusCodes.OK);
-    expect(json.length).toBe(2);
-    expect(json[0].serviceId).toBe(SERVICE_ID_1);
-    expect(json[1].serviceId).toBe(SERVICE_ID_2);
-  });
-
-  it('should return schemas filtered by platform codes', async () => {
-    const platformCodes = ['POSTGRESQL'];
-    const service = {
-      id: 'bd7a4c7a-1234-4c5e-b123-df12345abcd1',
-      name: 'User Service',
-      platformCode: platformCodes[0],
-    }
-
-    const filteredSchemas = [
-      {
-        id: 1,
-        name: "products",
-        description: "Table for product info",
-        schemaText: "CREATE TABLE products (id SERIAL PRIMARY KEY);",
-        serviceId: 'bd7a4c7a-1234-4c5e-b123-df12345abcd1',
-        service: service,
-      },
-      {
-        id: 2,
-        name: "users",
-        description: "Table for user info",
-        schemaText: "CREATE TABLE users (id SERIAL PRIMARY KEY);",
-        serviceId: 'bd7a4c7a-1234-4c5e-b123-df12345abcd2',
-        service: service,
-      },
-    ];
-
-    (prisma.schema.findMany as jest.Mock).mockResolvedValue(filteredSchemas);
-
-    const request = new NextRequest(
-      new URL(
-        `http://localhost/api/schema?platformCodes=${platformCodes[0]}`
-      ),
-      {
-        method: "GET",
-      }
-    );
-
-    const response = await GET(request);
-    const json = await response.json();
-
-    expect(prisma.schema.findMany).toHaveBeenCalledWith(expect.objectContaining({
-      where: {
-        AND:
-          expect.arrayContaining([
-            { service: { platformCode: expect.objectContaining({ in: platformCodes }) } }
-          ]),
-      }
-    }));
-
-    expect(response.status).toBe(StatusCodes.OK);
-    expect(json.length).toBe(2);
-    expect(json[0].service.platformCode).toBe(platformCodes[0]);
-    expect(json[1].service.platformCode).toBe(platformCodes[0]);
-  })
-
-  it('should return response 400 Bad Request for invalid platform codes', async () => {
-    const invalidPlatformCodes = ['INVALID_PLATFORM'];
-
-    const request = new NextRequest(
-      new URL(
-        `http://localhost/api/schema?platformCodes=${invalidPlatformCodes[0]}`
-      ),
-      {
-        method: "GET",
-      }
-    );
-
-    const response = await GET(request);
-
-    expect(response.status).toBe(StatusCodes.BAD_REQUEST);
-  })
-
-  it("should return 400 Bad Request for serviceIds other than UUID", async () => {
-    const invalidServiceId = "invalid-service-id";
-
-    const request = new NextRequest(
-      new URL(
-        `http://localhost/api/schema?serviceIds=${invalidServiceId}`
-      ),
-      {
-        method: "GET",
-      }
-    );
-
-    const response = await GET(request);
-    const json = await response.json();
-
-    expect(response.status).toBe(StatusCodes.BAD_REQUEST);
-    expect(json.message).toBe("Service ID must be a valid UUID");
-  })
 
 });
